+++ app/src/main/java/woowacourse/movie/view/MovieListActivity.kt
@@ -7,6 +7,7 @@ import androidx.appcompat.app.AppCompatActivity
 import woowacourse.movie.R
 import com.example.domain.Movie
 import woowacourse.movie.repository.MovieMockRepository
+import woowacourse.movie.view.model.toUiModel
 
 class MovieListActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
