+++ app/src/main/java/woowacourse/movie/view/ReservationActivity.kt
@@ -9,9 +9,12 @@ import woowacourse.movie.R
 import woowacourse.movie.databinding.ActivityReservationBinding
 import com.example.domain.Reservation
 import com.example.domain.ScreeningTime
+import woowacourse.movie.util.DATE_FORMATTER
 import woowacourse.movie.util.getParcelableCompat
 import woowacourse.movie.util.getSerializableCompat
+import woowacourse.movie.view.model.ReservationOptions
 import woowacourse.movie.view.MovieListActivity.Companion.MOVIE_ITEM
+import woowacourse.movie.view.model.MovieUiModel
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.LocalTime
