+++ app/src/main/java/woowacourse/movie/view/model/ReservationUiModel.kt
@@ -1,8 +1,8 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.model
 
 import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
 import com.example.domain.Reservation
+import kotlinx.parcelize.Parcelize
 import java.time.LocalDateTime
 
 fun Reservation.toUiModel(): ReservationUiModel = ReservationUiModel(
