+++ app/src/main/java/woowacourse/movie/view/model/SeatUi.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.model
 
 import com.example.domain.Seat
